+++ tests/regression/finite_volume/combined/gold/result.pt
